requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-paypal==2.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
pytz==2024.1
requests==2.31.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
